+++ src/basic/components/ui/cart/ProductList.tsx
@@ -1,5 +1,5 @@
-import { ProductWithUI } from '../../types'
-import { formatPrice } from '../../utils/formatters'
+import { ProductWithUI } from '../../../types'
+import { formatPrice } from '../../../utils/formatters'
 
 export function ProductList({
   filteredProducts,
